+++ CHANGES.rst
@@ -11,6 +11,8 @@ https://zope.readthedocs.io/en/2.13/CHANGES.html
 Bugs Fixed
 ++++++++++
 
+- #116: Restore exception views for unauthorized.
+
 - Restore a `_unauthorized` hook on the response object.
 
 - Restore `HTTPResponse.redirect` behaviour of not raising an exception.